fixing workflow tests
