+++ src/Esp.Net/Esp.Net.csproj
@@ -73,12 +73,12 @@
     <Compile Include="ObservationStage.cs" />
     <Compile Include="ObserveBaseEventAttribute.cs" />
     <Compile Include="ObserveEventAttribute.cs" />
-    <Compile Include="Plugins\HeldEvents\HeldEventAction.cs" />
-    <Compile Include="Plugins\HeldEvents\HeldEventActionEvent.cs" />
-    <Compile Include="Plugins\HeldEvents\IEventDescription.cs" />
-    <Compile Include="Plugins\HeldEvents\IEventHoldingStrategy.cs" />
-    <Compile Include="Plugins\HeldEvents\IHeldEventStore.cs" />
-    <Compile Include="Plugins\HeldEvents\RouterExt.cs" />
+    <Compile Include="HeldEvents\HeldEventAction.cs" />
+    <Compile Include="HeldEvents\HeldEventActionEvent.cs" />
+    <Compile Include="HeldEvents\IEventDescription.cs" />
+    <Compile Include="HeldEvents\IEventHoldingStrategy.cs" />
+    <Compile Include="HeldEvents\IHeldEventStore.cs" />
+    <Compile Include="HeldEvents\RouterExt.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Reactive\EventObservable.cs" />
     <Compile Include="Reactive\EventObserver.cs" />
