+++ src/Esp.Net/HeldEvents/IEventHoldingStrategy.cs
@@ -15,12 +15,17 @@
 #endregion
 
 #if ESP_EXPERIMENTAL
-namespace Esp.Net.Plugins.HeldEvents
+namespace Esp.Net.HeldEvents
 {
-    public enum HeldEventAction
+    public interface IEventHoldingStrategy<in TModel, in TEvent> where TEvent : IIdentifiableEvent
     {
-        Release,
-        Ignore
+        bool ShouldHold(TModel model, TEvent @event, IEventContext context);
+        IEventDescription GetEventDescription(TModel model, TEvent @event);
+    }
+
+    public interface IEventHoldingStrategy<in TModel, in TEvent, in TBaseEvent> : IEventHoldingStrategy<TModel, TEvent>
+        where TEvent : IIdentifiableEvent, TBaseEvent
+    {
     }
 }
 #endif
