+++ src/Esp.Net/HeldEvents/IHeldEventStore.cs
@@ -1,4 +1,4 @@
-#region copyright
+﻿#region copyright
 // Copyright 2015 Keith Woods
 // 
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -15,21 +15,13 @@
 #endregion
 
 #if ESP_EXPERIMENTAL
-using System;
 
-namespace Esp.Net.Plugins.HeldEvents
+namespace Esp.Net.HeldEvents
 {
-    public class HeldEventActionEvent
+    public interface IHeldEventStore
     {
-        public HeldEventActionEvent(Guid eventId, HeldEventAction action)
-        {
-            Action = action;
-            EventId = eventId;
-        }
-
-        public HeldEventAction Action { get; private set; }
-
-        public Guid EventId { get; private set; }
+        void AddHeldEventDescription(IEventDescription description);
+        void RemoveHeldEventDescription(IEventDescription description);
     }
 }
 #endif
